landing page video added
