+++ frontend/src/pages/Dashboard.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
-import { signOut } from "firebase/auth";
-import { auth } from "../Config"; // Adjust path as needed
+import Sidebar from "../components/Sidebar"; // ✅ import the sidebar
 
 const Dashboard = () => {
   const navigate = useNavigate();
@@ -11,7 +10,7 @@ const Dashboard = () => {
       title: "Start Interview",
       description:
         "Begin a new AI-powered mock interview session and test your readiness in real-world scenarios.",
-      route: "/pre-interview", // ✅ updated to go to pre-interview
+      route: "/pre-interview",
       icon: "🗣️",
     },
     {
@@ -30,45 +29,10 @@ const Dashboard = () => {
     },
   ];
 
-  const handleLogout = async () => {
-    try {
-      await signOut(auth);
-      navigate("/login");
-      console.log("Logout Successful");
-    } catch (error) {
-      console.error("Error during sign-out:", error.message);
-      alert("Failed to log out. Please try again.");
-    }
-  };
-
   return (
     <div className="min-h-screen flex bg-gradient-to-b from-[#0f172a] to-[#1e293b] text-white">
-      {/* Sidebar */}
-      <aside className="w-64 bg-[#1e293b] p-6 flex flex-col justify-between fixed left-0 top-0 bottom-0 pt-24 shadow-lg">
-        <div className="flex flex-col items-center space-y-4">
-          <img
-            src=""
-            alt="Profile"
-            className="w-24 h-24 rounded-full border-4 border-yellow-400"
-          />
-          <h2 className="text-xl font-semibold text-yellow-400">Welcome</h2>
-          <p className="text-gray-400 text-sm">User Name</p>
-
-          <button
-            onClick={() => navigate("/")}
-            className="w-full mt-6 bg-gray-700 text-white px-4 py-2 rounded hover:bg-yellow-500 hover:text-black transition"
-          >
-            ⬅️ Home
-          </button>
-        </div>
-
-        <button
-          onClick={handleLogout}
-          className="bg-yellow-400 text-black px-4 py-2 rounded hover:bg-yellow-500 transition"
-        >
-          Logout
-        </button>
-      </aside>
+      {/* ✅ Reusable Sidebar */}
+      <Sidebar />
 
       {/* Main Content */}
       <main className="flex-1 ml-64 px-6 pt-24 pb-16">
